doc-analyzer/backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
openai==1.76.2
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.6.3 